Refactor: Rename and reorganize OpenAPI specification files and update paths
