+++ api/openapi.yml
@@ -8,7 +8,7 @@ servers:
   - url: http://localhost:3000
     description: Local server on port 3000
   - url: http://localhost:4000
-    description: Local server on port 4000
+    description: Local server on port 5000
   - url: https://gfaiers.com
     description: Production server
 paths:
