+++ api/src/helpers/spec.ts
@@ -4,7 +4,7 @@ import path from 'path'
 import fs from 'fs'
 import yaml from 'js-yaml'
 
-const swaggerPath = path.join(__dirname, 'openapi.yml')
+const swaggerPath = path.join(__dirname, '../../openapi.yml')
 const swaggerDocument = yaml.load(fs.readFileSync(swaggerPath, 'utf8')) as Record<string, any>
 
 export const setupSwagger = (app: Application): void => {
